requirements.txt
labelImg
pymongo
boto3
pillow
